student_management/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
bottle==0.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clr_loader==0.2.7.post0
crispy-bootstrap5==2024.10
distlib==0.3.9
Django==4.2.16
django-crispy-forms==2.3
django-filter==24.3
filelock==3.18.0
gunicorn==23.0.0
idna==3.10
lxml==5.4.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyMySQL==1.1.1
python-docx==1.1.2
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32-ctypes==0.2.3
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0